backend/requirements.txt
uvicorn
fastapi
PyJWT[crypto]>=2.0.0
python-dotenv
numpy
pydantic
boto3
stripe
psycopg2-binary
python-json-logger
backoff
bs4
yt_dlp
pytube
moviepy
sqlalchemy
pg8000
cloud-sql-python-connector[pg8000]
pydantic[email]
